+++ packages/suite/src/hooks/suite/index.ts
@@ -5,7 +5,7 @@ export { useAnalytics } from './useAnalytics';
 export { useActions } from './useActions';
 export { useLayoutSize } from './useLayoutSize';
 export { useGraph } from './useGraph';
-export { useDebounce } from './useDebounce';
+export { useAsyncDebounce } from './useAsyncDebounce';
 export { useAccountSearch } from './useAccountSearch';
 export { useFirmware } from './useFirmware';
 export { useSelector } from './useSelector';
